Suppress roku power off timeout errors (#67414)
